requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
openpyxl
python-pptx
pdfplumber
xlrd
SQLAlchemy
pymysql
python-dotenv==1.0.1

